requirements.txt
Flask
PyMuPDF
python-docx
nltk
spacy
numpy
transformers
torch
firebase-admin
reportlab
gunicorn
werkzeug
requests